Arrange 3
Rearrange array such that arr[i] >= arr[j] if i 
is even and arr[i]<=arr[j] if i is odd and j < i.
Given an array of n elements. Our task is to write a program to rearrange 
the array such that elements at even positions are greater than all 
elements before it and elements at odd positions are less than all elements before it.